Add a standard logger and HTTP logging middleware
Create a standard log object and use dependency injection to pass it to
the broadbandCollector and a new loggingMiddleware for wrapping HTTP
requests to log the request.

Logs are emitted as structured JSON logs by default but can be switched
to a human-friendly output by passing -log.level console.

The Prometheus client_golang library has also been upgraded.